Massive module structure refactoring
