+++ tests/test_analysis.py
@@ -19,10 +19,10 @@ import unittest
 from random import randint
 from urllib.request import urlretrieve
 
-from pyserini.analysis import pyanalysis
-from pyserini.index import pyutils
+from pyserini import analysis
+from pyserini import index
+from pyserini import search
 from pyserini.pyclass import JString, JAnalyzerUtils
-from pyserini.search import pysearch
 
 
 class TestAnalyzers(unittest.TestCase):
@@ -39,18 +39,18 @@ class TestAnalyzers(unittest.TestCase):
         tarball = tarfile.open(self.tarball_name)
         tarball.extractall(self.index_dir)
         tarball.close()
-        self.searcher = pysearch.SimpleSearcher(f'{self.index_dir}lucene-index.cacm')
-        self.index_utils = pyutils.IndexReaderUtils(f'{self.index_dir}lucene-index.cacm')
+        self.searcher = search.SimpleSearcher(f'{self.index_dir}lucene-index.cacm')
+        self.index_utils = index.IndexReaderUtils(f'{self.index_dir}lucene-index.cacm')
 
     def test_different_analyzers_are_different(self):
-        self.searcher.set_analyzer(pyanalysis.get_lucene_analyzer(stemming=False))
+        self.searcher.set_analyzer(analysis.get_lucene_analyzer(stemming=False))
         hits_first = self.searcher.search('information retrieval')
-        self.searcher.set_analyzer(pyanalysis.get_lucene_analyzer())
+        self.searcher.set_analyzer(analysis.get_lucene_analyzer())
         hits_second = self.searcher.search('information retrieval')
         self.assertNotEqual(hits_first, hits_second)
 
     def test_analyze_with_analyzer(self):
-        tokenizer = pyanalysis.get_lucene_analyzer(stemming=False)
+        tokenizer = analysis.get_lucene_analyzer(stemming=False)
         query = JString('information retrieval')
         only_tokenization = JAnalyzerUtils.analyze(tokenizer, query)
         token_list = []
@@ -60,39 +60,39 @@ class TestAnalyzers(unittest.TestCase):
 
     def test_analysis(self):
         # Default is Porter stemmer
-        analyzer = pyanalysis.Analyzer(pyanalysis.get_lucene_analyzer())
+        analyzer = analysis.Analyzer(analysis.get_lucene_analyzer())
         tokens = analyzer.analyze('City buses are running on time.')
         self.assertEqual(tokens, ['citi', 'buse', 'run', 'time'])
 
         # Specify Porter stemmer explicitly
-        analyzer = pyanalysis.Analyzer(pyanalysis.get_lucene_analyzer(stemmer='porter'))
+        analyzer = analysis.Analyzer(analysis.get_lucene_analyzer(stemmer='porter'))
         tokens = analyzer.analyze('City buses are running on time.')
         self.assertEqual(tokens, ['citi', 'buse', 'run', 'time'])
 
         # Specify Krovetz stemmer explicitly
-        analyzer = pyanalysis.Analyzer(pyanalysis.get_lucene_analyzer(stemmer='krovetz'))
+        analyzer = analysis.Analyzer(analysis.get_lucene_analyzer(stemmer='krovetz'))
         tokens = analyzer.analyze('City buses are running on time.')
         self.assertEqual(tokens, ['city', 'bus', 'running', 'time'])
 
         # No stemming
-        analyzer = pyanalysis.Analyzer(pyanalysis.get_lucene_analyzer(stemming=False))
+        analyzer = analysis.Analyzer(analysis.get_lucene_analyzer(stemming=False))
         tokens = analyzer.analyze('City buses are running on time.')
         self.assertEqual(tokens, ['city', 'buses', 'running', 'time'])
 
         # No stopword filter, no stemming
-        analyzer = pyanalysis.Analyzer(pyanalysis.get_lucene_analyzer(stemming=False, stopwords=False))
+        analyzer = analysis.Analyzer(analysis.get_lucene_analyzer(stemming=False, stopwords=False))
         tokens = analyzer.analyze('City buses are running on time.')
         self.assertEqual(tokens, ['city', 'buses', 'are', 'running', 'on', 'time'])
 
         # No stopword filter, with stemming
-        analyzer = pyanalysis.Analyzer(pyanalysis.get_lucene_analyzer(stemming=True, stopwords=False))
+        analyzer = analysis.Analyzer(analysis.get_lucene_analyzer(stemming=True, stopwords=False))
         tokens = analyzer.analyze('City buses are running on time.')
         self.assertEqual(tokens, ['citi', 'buse', 'ar', 'run', 'on', 'time'])
 
     def test_invalid_analysis(self):
         # Invalid configuration, make sure we get an exception.
         with self.assertRaises(ValueError):
-            pyanalysis.Analyzer(pyanalysis.get_lucene_analyzer('blah'))
+            analysis.Analyzer(analysis.get_lucene_analyzer('blah'))
 
     def tearDown(self):
         self.searcher.close()
